update: update function delete , find account active = 1 , no activate = 0 return client

+++ module/account/biz/delete_account.go
@@ -1,6 +1,7 @@
 package accountbiz
 
 import (
+	"Ecommerce_Golang/common"
 	accountmodel "Ecommerce_Golang/module/account/model"
 	"context"
 )
@@ -15,17 +16,19 @@ type deleteAccountBiz struct {
 	store DeleteAccountStore
 }
 
-func NewDeleteAccount(store DeleteAccountStore) *deleteAccountBiz {
+func NewDeleteAccountBiz(store DeleteAccountStore) *deleteAccountBiz {
 	return &deleteAccountBiz{store: store}
 }
 
 func (biz *deleteAccountBiz) DeleteAccount(ctx context.Context, id int) error {
 
-	//var oldData *accountmodel.Account := biz.store.FindAccountWithCondition(ctx, map[string]interface{}{"id": id})
-	//
-	//if oldData.Status  == 0{
-	//	return errors.New("data has been deleted")
-	//}
+	oldData, err := biz.store.FindAccountWithCondition(ctx, map[string]interface{}{"id": id})
+	if err != nil {
+		return common.ErrEntityNotFound(common.EntityName, err)
+	}
+	if oldData.Status == 0 {
+		return common.ErrEntityDeleted(common.EntityName, nil)
+	}
 
 	if err := biz.store.Delete(ctx, id); err != nil {
 		return err
